add new inquiry feature completed

+++ src/main/java/com/jb/fop/dto/DashboardResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class DashboardResponse {
-    private Integer totalInquiryCount;
-    private Integer enrolledCount;
-    private Integer lostCount;
+    private long totalInquiryCount;
+    private long enrolledCount;
+    private long lostCount;
 }
